+++ server/out/server.js
@@ -13,60 +13,25 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const vscode_languageserver_1 = require("vscode-languageserver");
+const idl_routine_helper_1 = require("./providers/idl-routine-helper");
+const idl_document_symbol_manager_1 = require("./providers/idl-document-symbol-manager");
+const IDL_MODE = { language: "idl", scheme: "file" };
 // Create a connection for the server. The connection uses Node's IPC as a transport.
 // Also include all preview / proposed LSP features.
 let connection = vscode_languageserver_1.createConnection(vscode_languageserver_1.ProposedFeatures.all);
 // Create a simple text document manager. The text document manager
 // supports full document sync only
 let documents = new vscode_languageserver_1.TextDocuments();
+// create all of our helper objects for different requests
+const routineHelper = new idl_routine_helper_1.IDLRoutineHelper(connection, documents);
+const symbolProvider = new idl_document_symbol_manager_1.IDLDocumentSymbolManager(connection, documents);
+// flags for configuration
 let hasConfigurationCapability = false;
-let hasWorkspaceFolderCapability = false;
+let hasWorkspaceFolderCapability = true;
 let hasDiagnosticRelatedInformationCapability = false;
-// load our different routines
-const idlRoutines = require("../routines/idl.json");
-// give proper symbol
-idlRoutines.docs.forEach((item, idx) => {
-    // get key as string
-    const str = idx.toString();
-    // check for our system variable !null
-    if (item.label === null) {
-        item.label = "!null";
-    }
-    // handle setting proper information for our data things and such
-    switch (true) {
-        case idlRoutines.functions[str]:
-            item.insertText = item.label + "(";
-            item.kind = vscode_languageserver_1.CompletionItemKind.Function;
-            break;
-        case idlRoutines.procedures[str]:
-            item.insertText = item.label + ",";
-            item.kind = vscode_languageserver_1.CompletionItemKind.Function;
-            break;
-        case item.label.startsWith("!"):
-            item.kind = vscode_languageserver_1.CompletionItemKind.Constant;
-            break;
-        default:
-            item.kind = vscode_languageserver_1.CompletionItemKind.Text;
-    }
-    // check if we are an ENVI task, replace with ENVITask('TaskName')
-    if (item.label.startsWith("ENVI") && item.label.endsWith("Task")) {
-        item.insertText =
-            "ENVITask('" +
-                item.label.substr(0, item.label.length - 4).substr(4) +
-                "')";
-    }
-    // check if we are an IDL task, replace with ENVITask('TaskName')
-    if (item.label.startsWith("IDL") && item.label.endsWith("Task")) {
-        item.insertText =
-            "IDLTask('" +
-                item.label.substr(0, item.label.length - 4).substr(3) +
-                "')";
-    }
-    // save change
-    idlRoutines.docs[idx] = item;
-});
 connection.onInitialize((params) => {
     let capabilities = params.capabilities;
+    // params.workspaceFolders.
     // Does the client support the `workspace/configuration` request?
     // If not, we will fall back using global settings
     hasConfigurationCapability = !!(capabilities.workspace && !!capabilities.workspace.configuration);
@@ -80,21 +45,40 @@ connection.onInitialize((params) => {
             // Tell the client that the server supports code completion
             completionProvider: {
                 resolveProvider: true
-            }
+            },
+            definitionProvider: true,
+            workspaceSymbolProvider: true,
+            documentSymbolProvider: true
         }
     };
 });
-connection.onInitialized(() => {
+connection.onInitialized(() => __awaiter(this, void 0, void 0, function* () {
     if (hasConfigurationCapability) {
         // Register for all configuration changes.
         connection.client.register(vscode_languageserver_1.DidChangeConfigurationNotification.type, undefined);
     }
+    // listen for workspace folder event changes and update our serve-side cache
+    // TODO: detect when workspace is closed and remove files
     if (hasWorkspaceFolderCapability) {
-        connection.workspace.onDidChangeWorkspaceFolders(_event => {
-            connection.console.log("Workspace folder change event received.");
+        // get the list of current workspaces
+        connection.workspace.getWorkspaceFolders().then(folders => {
+            // when we have our folder, index files which should have all files?
+            symbolProvider.indexWorkspaces(folders).catch(err => {
+                connection.console.log(JSON.stringify(err));
+            });
         });
+        connection.workspace.connection.workspace // listen for new workspaces
+            .onDidChangeWorkspaceFolders((_event) => __awaiter(this, void 0, void 0, function* () {
+            connection.console.log("Workspace folder change event received. " + JSON.stringify(_event));
+            // refresh our index in case we have additional files
+            // because we cache results, should only be incremental processing
+            // as we add more folders
+            yield symbolProvider.indexWorkspaces(_event.added).catch(err => {
+                connection.console.log(JSON.stringify(err));
+            });
+        }));
     }
-});
+}));
 // The global settings, used when the `workspace/configuration` request is not supported by the client.
 // Please note that this is not the case when using this server with the client provided in this example
 // but could happen with other clients.
@@ -131,11 +115,16 @@ function getDocumentSettings(resource) {
 documents.onDidClose(e => {
     documentSettings.delete(e.document.uri);
 });
+// TODO: work with just the changed parts of a document
 // The content of a text document has changed. This event is emitted
 // when the text document first opened or when its content has changed.
-documents.onDidChangeContent(change => {
-    validateTextDocument(change.document);
-});
+documents.onDidChangeContent((change) => __awaiter(this, void 0, void 0, function* () {
+    // generate new symbols with the update, seems to magically sync when there are changes?
+    const newSymbols = yield symbolProvider.update(change.document.uri);
+}));
+documents.onDidOpen((event) => __awaiter(this, void 0, void 0, function* () {
+    yield symbolProvider.get.documentSymbols(event.document.uri);
+}));
 function validateTextDocument(textDocument) {
     return __awaiter(this, void 0, void 0, function* () {
         // In this simple example we get the settings for every validate run.
@@ -187,23 +176,32 @@ connection.onDidChangeWatchedFiles(_change => {
 });
 // This handler provides the initial list of the completion items.
 connection.onCompletion((_textDocumentPosition) => {
-    // The pass parameter contains the position of the text document in
-    // which code complete got requested. For the example we ignore this
-    // info and always provide the same completion items.
-    return idlRoutines.docs;
+    return routineHelper.completion(_textDocumentPosition);
 });
 // when we auto complete, do any custom adjustments to the data before the auto-complete
 // request gets back to the client
 connection.onCompletionResolve((item) => {
-    // // get the id
-    // const key = item.data.toString();
-    // // check if function or procedure
-    // switch (true) {
-    // 	default:
-    // 		// do nothing
-    // }
-    return item;
+    return routineHelper.postCompletion(item);
 });
+// handle when a user searches for a symbol
+connection.onWorkspaceSymbol((params) => {
+    return symbolProvider.searchByName(params.query);
+});
+// handle when we want the definition of a symbol
+connection.onDefinition((params) => {
+    const res = symbolProvider.searchByLine(params);
+    return res;
+});
+// connection.onHover(
+//   (params: TextDocumentPositionParams): Hover => {
+//     const res = symbolProvider.searchByLine(params);
+//     return res;
+//   }
+// )
+// handle when we request document symbols
+connection.onDocumentSymbol((params) => __awaiter(this, void 0, void 0, function* () {
+    return yield symbolProvider.get.documentSymbols(params.textDocument.uri);
+}));
 /*
 connection.onDidOpenTextDocument((params) => {
     // A text document got opened in VSCode.